Added .move and @two_ranks for Pawn
The possibilities of movement are based on instance variables for each
direction/rule. The .move will create a list of possible moves from
current position (disregarding legality) and enable the new position if
the move is within range.

It may be better to simply return a boolean for exiting movement