Piece and ValidMoves Classes Created [WIP]
The class objects need to be completed. The repetitive nature of the
previous piece objects for each piece can simplified into a hash lookup
for the only unique characteristics of their name and movement
rules (pattern).

Now each piece has an id in the hash containing their rules which are
used on initialization to create their unique traits.

However, due to the nature of tests, many of them will fail until we
finish the valid moves checker to recreate the same effect as the
previous methods had. The tests will only require a minor change of
passing the name of the piece to a Piece object in their `let` creation.